requirements.txt
selenium
webdriver-manager
beautifulsoup4
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
pyyaml